refactor: rework building the adjacency map
